+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 - implement a qsf data repository manager
 - ElasticWriter: add a alias at the beginning of the feeding, if no alias exists
 - read variables from a settings file in the home directory
+- allow classpath references in IOUTILS
 
 ### Changed